requirements.txt
-e .
streamlit
crewai
openai
python-dotenv
chromadb
crewai-tools
